commit #24

+++ Hiraeth/src/game/character_stats.h
@@ -27,7 +27,7 @@ namespace hiraeth {
 			unsigned int WeaponDef, MagicDef;
 			unsigned int WeaponAcc, MagicAcc;
 			unsigned int WeaponAvd, MagicAvd;
-			unsigned int Speed, Jump;
+			float Speed, Jump;
 		};
 
 		class CharacterStats : public Stats
